Move checkpoint creation and pruning into `Checkpoints`
